Thing Info panel: thing action is no longer shown in Doom map format. Sector and Thing Info panels: fixed possible crash when trying to show unknown thing/sector flag. Hints were not shown properly when multiselection was started while highlighting a map element. Copy/Paste actions work much faster now. Classic modes: rendered grid size in now adjusted to current zoom level.

+++ Source/Core/IO/UniversalStreamWriter.cs
@@ -290,7 +290,7 @@ namespace CodeImp.DoomBuilder.IO
 				if(t.Tag != 0) coll.Add("id", t.Tag);
 				coll.Add("x", t.Position.x);
 				coll.Add("y", t.Position.y);
-				if(t.Position.z != 0.0f) coll.Add("height", (float)t.Position.z);
+				if(t.Position.z != 0.0f) coll.Add("height", t.Position.z);
 				coll.Add("angle", t.AngleDoom);
 				coll.Add("type", t.Type);
 				if(t.Action != 0) coll.Add("special", t.Action);
